refactor: move Employment component
- Move the file src/components/Employment.tsx to
  src/components/employment/employment.tsx
- Update import paths to the Summary and Details components for the
  Employment component
- Update import paths to the Employment component for the Resume and
  Preview components
