+++ src/Resume.tsx
@@ -1,6 +1,6 @@
 import { Certificates } from "./components/Certificates";
 import { Education } from "./components/education/education";
-import { Employment } from "./components/Employment";
+import { Employment } from "./components/employment/employment";
 import { Interests } from "./components/interests/interests";
 import { PersonalDetails } from "./components/personal-details/personal-details";
 import { Qualifications } from "./components/qualifications/qualifications";
